+++ R/polmineR.R
@@ -50,7 +50,7 @@ setOldClass("quosure")
 #' 
 #' @section Package options:
 #' \itemize{
-#'   \item{\emph{polmineR.p_attribute}:} {The default attribute}
+#'   \item{\emph{polmineR.p_attribute}:} {The default positional attribute}
 #'   \item{\emph{polmineR.left}:} {Default value for left context.}
 #'   \item{\emph{polmineR.lineview}:} {A logical value, whether ...}
 #'   \item{\emph{polmineR.pagelength}:} {10L}
@@ -61,6 +61,10 @@ setOldClass("quosure")
 #'   \item{\emph{polmineR.buttons}:} {}
 #'   \item{\emph{polmineR.specialChars}:} {}
 #'   \item{\emph{polmineR.cutoff}:} {}
+#'   \item{\emph{polmineR.mdsub}:}{A list of pairs of character vectors
+#'   defining regular expression substitutions applied as part of preprocessing
+#'   documents for html display. Intended usage: Remove characters that would be
+#'   misinterpreted as markdown formatting instructions.}
 #'   \item{\emph{polmineR.corpus_registry}:} {The system corpus registry
 #'   directory defined by the environment variable CORPUS_REGISTRY before the
 #'   polmineR package has been loaded. The polmineR package uses a temporary